fix model and table

+++ database/migrations/2022_01_25_152709_create_travels_table.php
@@ -19,7 +19,7 @@ class CreateTravelsTable extends Migration
             $table->string('name');
             $table->string('place');
             $table->text('description');
-            $table->smallInteger('price');
+            $table->bigInteger('price');
         });
     }
